+++ example/index.ts
@@ -1,6 +1,6 @@
-import { Elysia, file } from 'elysia'
+import { Elysia, file, sse } from 'elysia'
 import { cors } from '@elysiajs/cors'
-import { swagger } from '@elysiajs/swagger'
+import { openapi } from '@elysiajs/openapi'
 
 import { node } from '../src'
 
@@ -8,24 +8,31 @@ const app = new Elysia({
 	adapter: node()
 })
 	.use(cors())
-	.use(swagger())
+	.use(openapi())
 	.ws('/ws/:id', {
-		open({ data }) {
-			console.log(data.params)
+		open({ data, subscribe, isSubscribed }) {
+			subscribe('welcome')
 		},
 		message(ws, message) {
 			ws.send(message)
-		}
+		},
+		close(ws, code, reason) {}
 	})
 	.get('/image', async () => file('test/kyuukurarin.mp4'))
 	.get('/generator', async function* () {
 		for (let i = 0; i < 100; i++) {
 			await new Promise((resolve) => setTimeout(resolve, 10))
-			yield 'A'
+			yield sse('A')
 		}
 	})
 	.post('/', ({ body }) => body, {
 		type: 'json'
 	})
-	.get('/', () => 'ok')
+	.get('/', ({ request }) => {
+		console.log(request)
+
+		return 'ok'
+	})
 	.listen(3000)
+
+// console.log(app.fetch.toString())